Support font palette overrides through SkFontArguments
Similar to how we allow configuration of variable font configurations,
provide additional SkFontArguments to select a base palette and a set
of potentially sparse color overrides.

This is required for implementing CSS font-palette.

Modify the more_samples-glyf_colr_1.ttf to have two additional palettes,
and two additional test glyphs, one that draws with COLRv0 logic, one
that draws with COLRv1 logic and has a foreground palette index dot
in the middle. See [1] & [2] for the additions to the test font.

Add a GM which tests this on the SkFontMgr_custom using makeClone()
and makeFromStreamArgs(). Implementations for Fuchsia and
SkFontMgr_FontConfig are currently untested. The test displays
the two glyphs in default palette on the left, then with palette
overrides (as in the title of the test) on the right. The first row
uses a typeface created with makeFromStreamArgs(), the second
uses one created with makeClone().

Change the Scanner in the FreeType backend to not only retrieve
axis definitions but also the palette information for a particular
requested basePalette from the SkFontArguments. Add a static method
that can resolve the rest of the font arguments against this
retrieved palette.

[1] https://github.com/googlefonts/color-fonts/pull/91
[2] https://github.com/googlefonts/color-fonts/pull/92

Bug: skia:12730, chromium:1170794
Cq-Include-Trybots: luci.skia.skia.primary:Test-Android-Clang-GalaxyS20-GPU-MaliG77-arm64-Release-All-Android_NativeFonts,Test-Debian10-Clang-GCE-CPU-AVX2-x86_64-Debug-All-NativeFonts
Change-Id: Ia1334f069240edc78fd4791969914e8a6f4fbaf9